python_scripts/requirements.txt
# ======================================
# Bibliotecas para Manipulação de Dados
# ======================================
pandas>=2.1.0
numpy>=1.24.0
openpyxl>=3.1.0

# ======================================
# Processamento de Documentos
# ======================================
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.0
reportlab>=4.0.0

# ======================================
# APIs e Web
# ======================================
requests>=2.31.0
urllib3>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0

# ======================================
# Geoprocessamento (para módulos futuros)
# ======================================
# geopandas>=0.14.0
# shapely>=2.0.0
# folium>=0.15.0
# pyproj>=3.6.0

# ======================================
# Cálculos de Engenharia
# ======================================
scipy>=1.11.0
sympy>=1.12

# ======================================
# IA e LLM
# ======================================
openai>=1.3.0
langchain>=0.1.0
langchain-openai>=0.0.2

# ======================================
# Utilidades
# ======================================
python-dotenv>=1.0.0
pydantic>=2.5.0
jinja2>=3.1.2
pyyaml>=6.0.1
toml>=0.10.2

# ======================================
# Data/Hora
# ======================================
python-dateutil>=2.8.2
pytz>=2023.3

# ======================================
# Logging e Monitoramento
# ======================================
loguru>=0.7.2

# ======================================
# Testes (opcional)
# ======================================
# pytest>=7.4.0
# pytest-cov>=4.1.0
